// YOUCOM.ReserVook/src/app/features/master/loststate/model/loststate.model.ts
export class LostStateInfo {
  companyNo: string;
  itemStateCode: string;
  itemStateName: string;
  color: string;
  defaultFlagSearch: string;
  defaultFlagEntry: string;
  orderNo: number;
  status: string;
  version: number;
  creator: string;
  updator: string;
  cdt: string;
  udt: string;
}